% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadExpr.R
\name{getGdcClient}
\alias{getGdcClient}
\title{Obtain the gdc-client tool path}
\usage{
getGdcClient()
}
\value{
the gdc-clent path
}
\description{
Obtain the gdc-client tool path
}
\examples{
\dontrun{
getGdcClient()
}
}
