% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadExpr.R
\name{installTools}
\alias{installTools}
\title{Install gdc-client tool}
\usage{
installTools(
 
    gdcClientURL = "https://gdc.cancer.gov/files/public/file/gdc-client_v1.6.1_OSX_x64.zip",
  manuallyDownloadFile = NULL
)
}
\arguments{
\item{gdcClientURL}{The URL for downloading gdc-client tool}

\item{manuallyDownloadFile}{The path of manually downloaded gdc-client .zip file}
}
\value{
The full path of downloaded/installed gdc-client tool.
}
\description{
Install gdc-client tool
}
\examples{
\dontrun{
# Download gdc-client from https://gdc.cancer.gov/access-data/gdc-data-transfer-tool.

# By default install gdc-client of macOS version
installTools()

# Install gdc-client of Linux (Ubuntu) version
installTools("https://gdc.cancer.gov/files/public/file/gdc-client_v1.6.1_Ubuntu_x64.zip")

# Install gdc-client of Windows version
installTools("https://gdc.cancer.gov/files/public/file/gdc-client_v1.6.1_Windows_x64.zip")
}
}
