% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles.R
\name{read_htseqCountsJSON}
\alias{read_htseqCountsJSON}
\title{Read htseq.counts JSON data}
\usage{
read_htseqCountsJSON(
  file = system.file("extdata/gdc_htseq_counts_LUAD_all.json", package = "TCGAnalysis")
)
}
\description{
Read htseq.counts JSON data
}
