% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadExpr.R
\name{readTCGA}
\alias{readTCGA}
\title{Read data downloaded from TCGA GDC repository}
\usage{
readTCGA(downloadFolder, dataType = c("htseq.counts", "clinical"), ...)
}
\description{
Read data downloaded from TCGA GDC repository
}
\examples{
\dontrun{
expr = readTCGA("~/Downloads/LUAD_expr5/htseq_counts/", dataType = "htseq.counts")
head(expr)
clinical = readTCGA("~/Downloads/LUAD_expr5/clinical/", dataType = "clinical")
head(clinical)
}
}
